feat: separate components for tin
