+++ packages/account/src/Components/forms/form-fields/index.ts
@@ -1,5 +1,7 @@
 import FormInputField from './form-input-field';
 import DateOfBirthField from './date-of-birth-field';
 import EmploymentStatusField from './employment-status';
+import TaxResidenceField from './tax-residence';
+import TaxIdentificationNumberField from './tax-indentification-number';
 
-export { FormInputField, DateOfBirthField, EmploymentStatusField };
+export { FormInputField, DateOfBirthField, EmploymentStatusField, TaxResidenceField, TaxIdentificationNumberField };
